feat: quiet option for console convert
There's a known issue with `System.Console` where it throws a "handle is invalid" error when manipulating the cursor outside of a "real" terminal session. I can confirm that this makes it impossible to call Repackinator from Python, and almost certainly in other languages as well.

Adding a `quiet` arg is a bit of a brute-force workaround, but it seems like a good idea to have the option in any case.